+++ application/app/views/changes.md
@@ -4,6 +4,7 @@
      * when marking as _read_ flush the cache
      * seen flag is read each parse, so if another client marks as seen, we also see
    * fix issue if two users were cleaning up cache simultaneously
+2. New Quick Reply feature
 
 ### 06/02/2020
 1. Fixed bug introduced yesterday - message images were not being indexed properly